Limited the #cores in PCA denoising to 75% to make PC/Laptop usuable during.
Added comments to decorrelation code and updated the mask-size of ncov_usingData to be equal to default mask size of noise-creation. Storing noise from voxel-data prior to noise-decor and using this "measurement" for noise-cov is equal to using voxel-data noise during noise-decor calculations.

Updated SNR_AUC - needs an update for multi-file use - tbd.